other: ヘルパー関数をテスト用ファイルに移動し、不要なコードを削除
